+++ cathode/src/main/java/net/simonvt/cathode/ui/adapter/ShowDescriptionAdapter.java
@@ -92,7 +92,7 @@ public class ShowDescriptionAdapter
     final boolean inCollection =
         cursor.getInt(cursor.getColumnIndex(ShowColumns.IN_COLLECTION_COUNT)) > 1;
     final boolean inWatchlist = cursor.getInt(cursor.getColumnIndex(ShowColumns.IN_WATCHLIST)) == 1;
-    final int rating = cursor.getInt(cursor.getColumnIndex(ShowColumns.RATING));
+    final float rating = cursor.getFloat(cursor.getColumnIndex(ShowColumns.RATING));
 
     holder.indicator.setWatched(watched);
     holder.indicator.setCollected(inCollection);
